requirements.txt
aniso8601==8.0.0
apispec==5.1.1
click==7.1.2
Flask==1.1.2
flask-apispec==0.11.0
Flask-JWT-Extended==3.24.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.14.1
packaging==21.3
passlib==1.7.2
PyJWT==1.7.1
pyparsing==3.0.7
python-decouple==3.8
pytz==2020.1
six==1.14.0
SQLAlchemy==1.3.16
webargs==8.1.0
Werkzeug==1.0.1
python-dotenv
psycopg2
aws-cdk-lib==2.27.0
flask_cors
boto3
python-ulid